lần này sẽ thành công/backend/requirements.txt
flask
flask-cors
opencv-python
pillow
numpy
gunicorn
